Reorganize commands
